revise for quitruple

+++ model_rule_rhs.py
@@ -8,13 +8,14 @@ from typing import Tuple, List, Dict
 import numpy as np
 import random
 import pandas as pd
+import os
 
 parser = argparse.ArgumentParser(
     description="Temporal ComplEx"
 )
 
 parser.add_argument(
-    '--dataset', type=str, default='ICEWS14RR',
+    '--dataset', type=str, default='YAGO',
     help="Dataset name"
 )
 
@@ -121,7 +122,10 @@ def find_reverse(instance):
 
 
 def find_freq_entity_rel(instance):
-    head, rel, tail, time = instance
+    if len(instance) == 4:
+        head, rel, tail, time = instance
+    else:
+        head, rel, tail, time, time_e = instance
     selected = (train[:,1] == rel) & (train[:, 0] == head)
     selected = train[selected]
     if len(selected) > 0:
@@ -133,7 +137,10 @@ def find_freq_entity_rel(instance):
 
 
 def find_freq_entity(instance):
-    head, rel, tail, time = instance
+    if len(instance) == 4:
+        head, rel, tail, time = instance
+    else:
+        head, rel, tail, time, time_e = instance
     selected = train[:, 0] == head
     selected = train[selected]
     if len(selected) > 0:
@@ -145,28 +152,33 @@ def find_freq_entity(instance):
 
 
 def make_dictionary(testset, if_relation=False):
-    dict = {}
-    if not if_relation:
-        for i, query in enumerate(testset):
-            print(if_relation, i)
-            dict[query[0]] = find_freq_entity(query)
-    else:
-        for i, query in enumerate(testset):
-            print(if_relation, i)
-            dict[(query[0], query[1])] = find_freq_entity_rel(query)
-    with open(f'dict_{args.dataset}_{if_relation}_rhs.pkl', 'wb') as f:
-        pickle.dump(dict, f)
+    if not os.path.exists(f'dict_{args.dataset}_{if_relation}_rhs.pkl'):
+        dict = {}
+        if not if_relation:
+            for i, query in enumerate(testset):
+                #print(if_relation, i)
+                dict[query[0]] = find_freq_entity(query)
+        else:
+            for i, query in enumerate(testset):
+                #print(if_relation, i)
+                dict[(query[0], query[1])] = find_freq_entity_rel(query)
+        with open(f'dict_{args.dataset}_{if_relation}_rhs.pkl', 'wb') as f:
+            pickle.dump(dict, f)
 
 
 def rank_entity_rules(instance, filter_out):
-    head, rel, tail, time = instance
+    if len(instance) == 4:
+        head, rel, tail, time = instance
+    else:
+        head, rel, tail, time, time_e = instance
     rank_entity, entity_temporal, entity_static = [], [], []
-    if rel in symmetry_rel.keys():
-        entity_temporal = find_symmetry_temporal(instance)
-        entity_static = find_symmetry(instance)
-    elif rel in reverse_rel.keys():
-        entity_temporal = find_reverse_temporal(instance)
-        entity_static = find_reverse(instance)
+    if args.dataset in ['ICEWS14RR', 'ICEWS15RR']:
+        if rel in symmetry_rel.keys():
+            entity_temporal = find_symmetry_temporal(instance)
+            entity_static = find_symmetry(instance)
+        elif rel in reverse_rel.keys():
+            entity_temporal = find_reverse_temporal(instance)
+            entity_static = find_reverse(instance)
     # freq_entity_rel = find_freq_entity_rel(instance)
     # freq_entity = find_freq_entity(instance)
 
@@ -210,7 +222,10 @@ def get_ranking(
     make_dictionary(queries, False)
     make_dictionary(queries, True)
     for i, query in enumerate(queries):
-        filter_out = filters[(query[0], query[1], query[3])]
+        if len(query) == 4:
+            filter_out = filters[(query[0], int(query[1]), query[3])]
+        else:
+            filter_out = filters[(query[0], int(query[1]), query[3], query[4])]
         rank_entity, rules = rank_entity_rules(query, filter_out)
         # filter_out = filters[(query[0], query[1], query[3])]
         # temp_dict = {'key': [query[0], query[1], query[3]], 'filter out': filter_out, 'label': [query[2]],
@@ -272,21 +287,21 @@ def hits_k_max(q, ranks):
         hits_at[m] = np.mean(max_ranks)
     return hits_at
 
-
-ranks = get_ranking().astype(int)
-macro = mean_rank_max(dataset.get_examples('test'), ranks)
-hits_macro = hits_k_max(dataset.get_examples('test'), ranks)
-print(f'Macro_rhs:{macro}')
-print(f'Macro_hits:{hits_macro}')
-print(ranks.mean())
-mean_reciprocal_rank = np.mean(1. / ranks)
-print(mean_reciprocal_rank)
-hits_at_1 = np.mean(ranks<=1)
-hits_at_3 = np.mean(ranks<=3)
-hits_at_10 = np.mean(ranks<=10)
-print(hits_at_1)
-print(hits_at_3)
-print(hits_at_10)
+if __name__ == '__main__':
+    ranks = get_ranking().astype(int)
+    macro = mean_rank_max(dataset.get_examples('test'), ranks)
+    hits_macro = hits_k_max(dataset.get_examples('test'), ranks)
+    print(f'Macro_rhs:{macro}')
+    print(f'Macro_hits:{hits_macro}')
+    print(ranks.mean())
+    mean_reciprocal_rank = np.mean(1. / ranks)
+    print(mean_reciprocal_rank)
+    hits_at_1 = np.mean(ranks<=1)
+    hits_at_3 = np.mean(ranks<=3)
+    hits_at_10 = np.mean(ranks<=10)
+    print(hits_at_1)
+    print(hits_at_3)
+    print(hits_at_10)
 
 # rhs_results = pd.DataFrame(columns=['Rank mean', 'MRR', 'hits 1', 'hits 3', 'hits 10'])
 # for _ in range(5):
